TPermissionsManager Unit Test added again, but minimal for testing on github
The TPermissionsManagerTest Unit test was added but it failed on almost all tests even though all the tests pass on my local machine.  So i removed the Test to see if that was the problem.  and everything else did pass.  So there is a problem in the unit test on the github platform to be debugged.